Forcing text encoding to UTF-8 and line endings to Unix-style in all OS
We have seen issues that one line changes of a file results in changes
of the whole file. While git offers some options to deal with this it
was not always applied properly. Hopefully this will improve in the
future by making Eclipse smart. Each sub project for Eclipse is now
configured with these settings to perform URF-8 encoding on the files
and Unix style line ending (\n), so that it should work fine even if the
git option autocrlf = input is not used properly.