New/rgroup decomposition (#72)
* Added thread synchronization to avoid NullPointerException raise cond.

Fixed line endings for Linux and encoding to UTF-8.

* Unfinished changes for new RGroupDecomposition node

* Undeprecated the old RDKit R-Group Decomposition node

This will give users the chance to try both versions, old and new, at
the same time, while the new version is still in beta state.

* Some improvements in InputDataInfo handling

Added new constructors to allow handover of a list of acceptable data
value classes to be used as input instead of using an array, which
caused issues when using generic data types.
Better handling for adapter cells for SMILES, SDF and SMARTS.

* Improved error handling for undefined enum selection parameter

* Added new RDKit R-Group Decomposition node as beta version.

This is based on RDKit functionality developed in 2018/2019. Eventually
this will completely replace the old RDKit R-Group Decomposition node,
which has been deprecated but can still be used in the moment.

* Renamed old R-Group Decomposition tests

* Added new temporary R-Group Decomposition tests

* Delete old test duplicate

* Updated copyright information

* Bugfixes and added Remove Empty Rx Columns parameter

* update expected results

* Bugfix to avoid RDKit exceptions after processing failure

* Updated R-Group Decomposition test workflow

* update expected results

* needs testing

* update expected test results

* update windows binary, remove old boost libs, update loader

* update tests for windows

* simplify RMSD tests to remove chances for roundoff error to bite us

* remove a bogus workflowset.meta file

* add target definition for KNIME 4.0

* bump the binaries to 4.0

* Added thread synchronization to avoid NullPointerException raise cond.

Fixed line endings for Linux and encoding to UTF-8.

* Unfinished changes for new RGroupDecomposition node

* Undeprecated the old RDKit R-Group Decomposition node

This will give users the chance to try both versions, old and new, at
the same time, while the new version is still in beta state.

* Some improvements in InputDataInfo handling

Added new constructors to allow handover of a list of acceptable data
value classes to be used as input instead of using an array, which
caused issues when using generic data types.
Better handling for adapter cells for SMILES, SDF and SMARTS.

* Improved error handling for undefined enum selection parameter

* Added new RDKit R-Group Decomposition node as beta version.

This is based on RDKit functionality developed in 2018/2019. Eventually
this will completely replace the old RDKit R-Group Decomposition node,
which has been deprecated but can still be used in the moment.

* Renamed old R-Group Decomposition tests

* Added new temporary R-Group Decomposition tests

* Delete old test duplicate

* Updated copyright information

* Bugfixes and added Remove Empty Rx Columns parameter

* Bugfix to avoid RDKit exceptions after processing failure

* Updated R-Group Decomposition test workflow

* add a smaller test with clear known results