Update requirements for KNIME's updated underlaying Eclipse version
From this commit on this package will work only with KNIME 4.7.0 and forward.

Apache Batik is now required in the currently latest version 1.14 instead of
the previous version 1.13. Import of org.w3c.dom.events was previously
required but would break the KNIME AP 4.7.0 and greater.

For details refer to
https://forum.knime.com/t/update-to-eclipse-2022-06-for-knime-ap-4-7-0/46059