Add RDKit type support for KNIME's new Python (Labs) integration. (#111)
* Add python3 folder to build.properties
* Add value type names for extension types
* Adjust deprecated import to avoid deprecation warnings
* Refactor serializers to remove interdependencies of Python
integrations

The "new" Python integration and the columnar backend use the
serialization as defined by the ValueFactories. These were re-using the
serialization methods from the Serializers and Deserializers for the
"old" Python integration, which unfortunately introduced a dependency to
the old Python integration.

This means up to now the old Python integration had to be installed for
the RDKit types to work in the "new" Python integration and the columnar
backend. We extract now the shared serialization methods into a utility
class, breaking the aforementioned dependency.