KNIME-1497: Allow strict parsing flag control for mol block parsing
There are several tools to sketch molecules. Sometimes, there are
certain violations of the official standard introduced by these tools,
and RDKit by default is very strict in parsing SDF structures. Hence,
users run from time to time into failures. There is a flag in RDKit
methods that allows to control how strict the parsing should be done. We
should investigate where this may have a good influence to control this
flag.

Investigation has shown that it makes sense to use in the following
scenarios:

* Auto-conversion: All RDKit nodes automatically turn SMILES or SDF
structures into RDKit molecules. The user cannot control this in the
moment. We should add a preference about strict parsing handling for
this case.
* Rendering: The RDKit 2D Depiction rendering takes also plain SDF
values and renders them after conversion into RDKit molecules. The user
cannot control the conversion in the moment. We should add a preference
about strict parsing handling for this case.
* Node Settings: Currently only 2 nodes could make use of the strict
parsing setting: The RDKit from Molecule node (for converting SDF
molecules) and the RDKit R Group Decomposition node (for reading cores
in SDF format). In the moment it always uses the strict parsing option.
We should add a node setting to control this. The default for old
existing nodes should always be true. For new nodes the default could be
controlled by a preference about strict parsing handling for this case.
In any case the user should be able to control the setting within the
node settings. As it makes only sense for handling SDF we should disable
it for all other molecule types.