KNIME-1499: Move preference init from start() method into initPrefs() (#124)
This bugfix ensures that the RDKit feature runs properly on KNIME
executors / servers. It addresses a bug that was introduced with the
last commit on Dec 20, 2022, when preference initialization code had
been added to a place that works only on KNIME AP (with GUI), but not in
executors (headless environments).