KNIME-1569: Improve structure rendering control (#127)
* KNIME-1569: Added option to use CoordGen for 2d coord generation

This setting can be controlled with a preference (default is the same as
before, using native RDKit rendering). It is mainly interesting for
structure rendering, but also influences some situations when
coordinates are being generated because no conformer was found yet in an
RDKit molecule.

The mol.compute2DCoords(smarts, ...) method is not available with all
other default values. This has been added implemented now and is called
for instance from the RDKitAddCoordinates node model. Without this
change, it would always use CoordGen (due to our new global setting) and
not honor the new preference for the CoordGen usage.

Algorithm was taken from
https://github.com/rdkit/rdkit/blob/58b79c6f8e2581205007effcf53c6c3641393c07/Code/JavaWrappers/ROMol.i#L441-L457
and converted into Java.