KNIME-1797: Updated Copyright NIBR to Novartis Pharma (#145)
* KNIME-1797: Updated NIBR => Novartis, Updated all Copyrights

The name Novartis Institutes for BioMedical Research was renamed. We
will start using Novartis Pharma AG, Switzerland, as the new legal
entity in the copyright. BioMedial Research is an internal sub devision.

* KNIME-1797: Remove all Java Doc from Git

Java Docs should have never been added to get, because they are not kept
up to date automatically, and if they would, we would have constant
changes, which takes unnecessary space.