KNIME-1797: Updated Copyright/Vendor from NIBR to Novartis Pharma
* Updated NIBR => Novartis, Updated all Copyrights / Vendor info
The name Novartis Institutes for BioMedical Research was renamed. We
will start using Novartis Pharma AG, Switzerland, as the new legal
entity in the copyright. BioMedial Research is an internal sub devision.

* Removed all Java Doc from Git
Java Docs should have never been added to git, because they are not kept
up to date automatically, and if they would, we would have constant
changes, which takes unnecessary space.

Added 5.2 launch configuration
Small fix in org.rdkit.knime.nodes.RDKitInteractiveView for KNIME 4.7 to
be compatible with 5.1 and 5.2