KNIME-1856: Adding Maven plugin to manage dependencies (#148)
As discussed in the KNIME RDKit Nodes Forum (https://forum.knime.com/t/adding-maven-dependencies-to-knime-node-plugin/77331) we wanted to have a way to update the MANIFEST.MF and build.properties file based on the latest fetched Maven dependencies that we specify in a special pom.xml file as demonstrated in this forum post.

Out of this need came a Maven plugin that performs this job. It is now used in RDKit Nodes (only OPSIN dependencies), in the Novartis BR Extension (with a lot more dependencies), and ready to be used by others in KNIME Community.